+++ srcs/utils/ft_realloc.c
@@ -6,16 +6,16 @@
 /*   By: hlee-sun <hlee-sun@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/16 13:26:48 by suminkwon         #+#    #+#             */
-/*   Updated: 2024/10/13 17:02:39 by hlee-sun         ###   ########.fr       */
+/*   Updated: 2024/10/13 19:28:32 by hlee-sun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-char	**ft_realloc_double(char **token, size_t position, size_t *bufsize)
+char	**ft_realloc_double(char **token, int position, int *bufsize)
 {
 	char	**realloc_token;
-	size_t		i;
+	int		i;
 
 	i = 0;
 	if (position >= *bufsize)
@@ -38,10 +38,10 @@ char	**ft_realloc_double(char **token, size_t position, size_t *bufsize)
 	return (token);
 }
 
-char *ft_realloc_single(char *token, size_t position, size_t *bufsize)
+char *ft_realloc_single(char *token, int position, int *bufsize)
 {
 	char	*realloc_token;
-	size_t		i;
+	int		i;
 
 	i = 0;
 	if (position >= *bufsize)
